+++ src/components/common/AdBanner.tsx
