+++ src/components/common/index.ts
@@ -1,0 +1,3 @@
+export { default as LoadingView } from './LoadingView';
+export { default as AdBanner } from './AdBanner';
+export { default as Error } from './Error';
