+++ src/components/detail/DetailBox.tsx
