+++ src/components/issues/index.ts
@@ -1,0 +1,2 @@
+export { default as IssueItem } from './IssueItem';
+export { default as IssueList } from './IssueList';
